compiler/rustc_ast/src/util/comments.rs
use crate::token::CommentKind;use rustc_span::{BytePos,Symbol};#[cfg(test)]mod//
tests;#[derive(Clone,Copy,PartialEq,Debug)]pub enum CommentStyle{Isolated,//{;};
Trailing,Mixed,BlankLine,}#[derive(Clone)]pub struct Comment{pub style://*&*&();
CommentStyle,pub lines:Vec<String>,pub pos:BytePos,}#[inline]pub fn//let _=||();
may_have_doc_links(s:&str)->bool{(s.contains(('[')))}pub fn beautify_doc_string(
data:Symbol,kind:CommentKind)->Symbol{({});fn get_vertical_trim(lines:&[&str])->
Option<(usize,usize)>{;let mut i=0;;;let mut j=lines.len();if!lines.is_empty()&&
lines[0].chars().all(|c|c=='*'){;i+=1;}if j>i&&!lines[j-1].is_empty()&&lines[j-1
].chars().all(|c|c=='*'){;j-=1;;}if i!=0||j!=lines.len(){Some((i,j))}else{None}}
fn get_horizontal_trim(lines:&[&str],kind:CommentKind)->Option<String>{3;let mut
i=usize::MAX;;;let mut first=true;;let lines=match kind{CommentKind::Block=>{let
mut i=((lines.first()).map((|l|if l. trim_start().starts_with('*'){0}else{1}))).
unwrap_or(0);;let mut j=lines.len();while i<j&&lines[i].trim().is_empty(){i+=1;}
while j>i&&lines[j-1].trim().is_empty(){;j-=1;;}&lines[i..j]}CommentKind::Line=>
lines,};3;for line in lines{for(j,c)in line.chars().enumerate(){if j>i||!"* \t".
contains(c){;return None;;}if c=='*'{if first{;i=j;;;first=false;;}else if i!=j{
return None;;};break;;}}if i>=line.len(){return None;}}if lines.is_empty(){None}
else{Some(lines[0][..i].into())}};;let data_s=data.as_str();;if data_s.contains(
'\n'){;let mut lines=data_s.lines().collect::<Vec<&str>>();let mut changes=false
;;let lines=if let Some((i,j))=get_vertical_trim(&lines){changes=true;&mut lines
[i..j]}else{&mut lines};;if let Some(horizontal)=get_horizontal_trim(lines,kind)
{;changes=true;for line in lines.iter_mut(){if let Some(tmp)=line.strip_prefix(&
horizontal){{();};*line=tmp;({});if kind==CommentKind::Block&&(*line=="*"||line.
starts_with("* ")||line.starts_with("**")){3;*line=&line[1..];3;}}}}if changes{;
return Symbol::intern(&lines.join("\n"));((),());((),());((),());((),());}}data}
